# perf: shapetracker ideas