@geohot 2024-09-14
This is adding complexity in exchange for speed. The other speed PRs were mostly complexity neutral.